* rename src/api to src/contracts (2/2)
moves all contract files into src/contracts directory
