+++ src/constants.ts
@@ -6,7 +6,7 @@ import {
   SDKOpts
 } from "./reach-helpers";
 import { PoolProtocolInfo } from "./types";
-import { getCurrentAnnouncers } from "./constants.versioned";
+import { announcersCurrent } from "./constants.versioned";
 
 // Strings
 export * from "./constants.strings";
@@ -149,7 +149,7 @@ function TriumvirContractId(opts: SDKOpts) {
   const { protocolId } = opts.contractOverrides || {};
   if (protocolId) return protocolId;
 
-  const { dev, mainnet } = getCurrentAnnouncers();
+  const { dev, mainnet } = announcersCurrent();
   return allocateProviderApps(opts.network, [
     dev.protocolId, // V3 Testnet Triumvirate
     mainnet.protocolId // V3 Mainnet Triumvirate
@@ -161,7 +161,7 @@ function LimitOrderAnnouncerId(opts: SDKOpts) {
   const { limitOrderAnnouncerId } = opts.contractOverrides || {};
   if (limitOrderAnnouncerId) return limitOrderAnnouncerId;
 
-  const { dev, mainnet } = getCurrentAnnouncers();
+  const { dev, mainnet } = announcersCurrent();
   return allocateProviderApps(opts.network, [
     dev.limitOrderAnnouncer, // V3 Testnet LO Announcer  (new)
     mainnet.limitOrderAnnouncer // V3 Mainnet LO Announcer  (new)
@@ -173,7 +173,7 @@ function PartnerFarmAnnouncerId(opts: SDKOpts) {
   const { partnerFarmAnnouncerId } = opts.contractOverrides || {};
   if (partnerFarmAnnouncerId) return partnerFarmAnnouncerId;
 
-  const { dev, mainnet } = getCurrentAnnouncers();
+  const { dev, mainnet } = announcersCurrent();
   return allocateProviderApps(opts.network, [
     dev.partnerFarmAnnouncerId, // V3 Testnet Partner Farms Announcer
     mainnet.partnerFarmAnnouncerId // V3 Mainnet Partner Farms Announcer
@@ -185,7 +185,7 @@ function PublicFarmAnnouncerId(opts: SDKOpts) {
   const { publicFarmAnnouncerId } = opts.contractOverrides || {};
   if (publicFarmAnnouncerId) return publicFarmAnnouncerId;
 
-  const { dev, mainnet } = getCurrentAnnouncers();
+  const { dev, mainnet } = announcersCurrent();
   return allocateProviderApps(opts.network, [
     dev.publicFarmAnnouncer, // V3 Testnet Public Farms Announcer
     mainnet.publicFarmAnnouncer // V3 Mainnet Public Farms Announcer
@@ -209,7 +209,7 @@ function allocateProviderApps(net?: NetworkProvider, opts: AppIds = [0, 0]) {
 
 /** @internal Get account address (not App ID!) of Triumvirate contract for current network */
 function ProtocolAddr(opts: SDKOpts) {
-  const currentversion = getCurrentAnnouncers();
+  const currentversion = announcersCurrent();
   const ADDRESSES = {
     MainNet: currentversion.mainnet.protocolAddress,
     None: "",
