+++ src/index.ts
@@ -19,15 +19,14 @@ export {
   setSlippage
 } from "./constants";
 
+export * from "./constants.versioned"
+
 // CONTRACTS
 export {
   LimitOrderType,
   LiquidityMigratorOpts,
   getLimitOrderVariant
 } from "./build/backend";
-
-// DATA & DATA FETCHERS
-export * from "./participants/index";
 
 // TYPES
 export * from "./types";
@@ -73,7 +72,7 @@ export {
   SDKOpts
 } from "./reach-helpers";
 
-// LIQUIDITY PROVIDER, SWAP, FARMING
+// LIQUIDITY PROVIDER, SWAP, FARMING, LIMIT ORDERS, etc
 export * from "./contracts/index";
 
 // UTILS | GENERAL
